+++ kitty/options/types.py
@@ -60,6 +60,7 @@ option_names = (  # {{{
  'background_tint',
  'bell_border_color',
  'bell_on_tab',
+ 'bell_path',
  'bold_font',
  'bold_italic_font',
  'box_drawing_scale',
@@ -457,6 +458,7 @@ class Options:
     background_tint: float = 0
     bell_border_color: Color = Color(red=255, green=90, blue=0)
     bell_on_tab: bool = True
+    bell_path: typing.Optional[str] = None
     bold_font: str = 'auto'
     bold_italic_font: str = 'auto'
     box_drawing_scale: typing.Tuple[float, float, float, float] = (0.001, 1.0, 1.5, 2.0)